refactor(eventForm): combine the creation and the edit  features in one form
KK-1017.
The event edit and create forms were already almost identical.
Now their features are combined in a 1 new event form
that both the views can use.
