+++ src/domain/events/ticketSystemInput/TicketSystemInput.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { SelectInput, SelectInputProps, useRecordContext } from 'react-admin';
+
+const TicketSystemInput = (props: SelectInputProps) => {
+  const record = useRecordContext();
+  return <SelectInput disabled={!!record?.publishedAt} {...props} />;
+};
+
+export default TicketSystemInput;
